pyproject.toml
[project]
name = "gemini-watermark-remover"
version = "0.1.0"
description = "Remove Gemini AI watermarks using reverse alpha blending"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name"}
]
keywords = ["gemini", "watermark", "image-processing", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pillow>=9.0.0",
    "numpy>=1.20.0",
    "flask>=3.0.0",
]

[project.scripts]
gemini-watermark-remover = "gemini_watermark_remover:main"

[project.urls]
Repository = "https://github.com/yourusername/gemini-watermark-remover"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
